ref(onboarding): Split setup docs files
