+++ static/app/views/onboarding/setUp/index.tsx
@@ -1,205 +1,39 @@
-import {Fragment, useCallback, useEffect, useMemo, useState} from 'react';
+import {Fragment, useCallback, useEffect, useState} from 'react';
 import {browserHistory} from 'react-router';
 import styled from '@emotion/styled';
 import {motion} from 'framer-motion';
-import {Location} from 'history';
 
 import {loadDocs} from 'sentry/actionCreators/projects';
-import {Alert} from 'sentry/components/alert';
-import ExternalLink from 'sentry/components/links/externalLink';
-import LoadingError from 'sentry/components/loadingError';
-import LoadingIndicator from 'sentry/components/loadingIndicator';
-import {DocumentationWrapper} from 'sentry/components/onboarding/documentationWrapper';
 import {Footer} from 'sentry/components/onboarding/footer';
 import {FooterWithViewSampleErrorButton} from 'sentry/components/onboarding/footerWithViewSampleErrorButton';
-import {PRODUCT, ProductSelection} from 'sentry/components/onboarding/productSelection';
 import {PlatformKey} from 'sentry/data/platformCategories';
 import platforms from 'sentry/data/platforms';
-import {t, tct} from 'sentry/locale';
+import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import {Organization, Project} from 'sentry/types';
 import {trackAnalytics} from 'sentry/utils/analytics';
-import getDynamicText from 'sentry/utils/getDynamicText';
 import {platformToIntegrationMap} from 'sentry/utils/integrationUtil';
-import {useApiQuery} from 'sentry/utils/queryClient';
 import useApi from 'sentry/utils/useApi';
 import {useExperiment} from 'sentry/utils/useExperiment';
 import useOrganization from 'sentry/utils/useOrganization';
+import {normalizeUrl} from 'sentry/utils/withDomainRequired';
 import SetupIntroduction from 'sentry/views/onboarding/components/setupIntroduction';
-import {SetupDocsLoader} from 'sentry/views/onboarding/setupDocsLoader';
+import {SdkLoader} from 'sentry/views/onboarding/setUp/sdkLoader';
 
-import FirstEventFooter from './components/firstEventFooter';
-import IntegrationSetup from './integrationSetup';
-import {StepProps} from './types';
-/**
- * The documentation will include the following string should it be missing the
- * verification example, which currently a lot of docs are.
- */
-const INCOMPLETE_DOC_FLAG = 'TODO-ADD-VERIFICATION-EXAMPLE';
+import FirstEventFooter from '../components/firstEventFooter';
+import IntegrationSetup from '../integrationSetup';
+import {StepProps} from '../types';
 
-type PlatformDoc = {html: string; link: string};
+import {SdkDoc} from './sdkDoc';
+import {SdkDocWithProductSelection} from './sdkDocWithProductSelection';
 
-function MissingExampleWarning({
-  platformDocs,
-  platform,
-}: {
-  platform: PlatformKey | null;
-  platformDocs: PlatformDoc | null;
-}) {
-  const missingExample = platformDocs?.html.includes(INCOMPLETE_DOC_FLAG);
+export type PlatformDoc = {html: string; link: string};
 
-  if (!missingExample) {
-    return null;
-  }
-
-  return (
-    <Alert type="warning" showIcon>
-      {tct(
-        `Looks like this getting started example is still undergoing some
-         work and doesn't include an example for triggering an event quite
-         yet. If you have trouble sending your first event be sure to consult
-         the [docsLink:full documentation] for [platform].`,
-        {
-          docsLink: <ExternalLink href={platformDocs?.link} />,
-          platform: platforms.find(p => p.id === platform)?.name,
-        }
-      )}
-    </Alert>
-  );
-}
-
-export function DocWithProductSelection({
-  organization,
+export function Setup({
+  route,
+  router,
   location,
-  projectSlug,
-  newOrg,
-  currentPlatform,
-}: {
-  currentPlatform: PlatformKey;
-  location: Location;
-  organization: Organization;
-  projectSlug: Project['slug'];
-  newOrg?: boolean;
-}) {
-  const loadPlatform = useMemo(() => {
-    const products = location.query.product ?? [];
-    return products.includes(PRODUCT.PERFORMANCE_MONITORING) &&
-      products.includes(PRODUCT.SESSION_REPLAY)
-      ? `${currentPlatform}-with-error-monitoring-performance-and-replay`
-      : products.includes(PRODUCT.PERFORMANCE_MONITORING)
-      ? `${currentPlatform}-with-error-monitoring-and-performance`
-      : products.includes(PRODUCT.SESSION_REPLAY)
-      ? `${currentPlatform}-with-error-monitoring-and-replay`
-      : `${currentPlatform}-with-error-monitoring`;
-  }, [location.query.product, currentPlatform]);
-
-  const {data, isLoading, isError, refetch} = useApiQuery<PlatformDoc>(
-    [`/projects/${organization.slug}/${projectSlug}/docs/${loadPlatform}/`],
-    {
-      staleTime: Infinity,
-      enabled: !!projectSlug && !!organization.slug && !!loadPlatform,
-    }
-  );
-
-  const platformName = platforms.find(p => p.id === currentPlatform)?.name ?? '';
-
-  return (
-    <Fragment>
-      {newOrg && (
-        <SetupIntroduction
-          stepHeaderText={t('Configure %s SDK', platformName)}
-          platform={currentPlatform}
-        />
-      )}
-      <ProductSelection
-        defaultSelectedProducts={[PRODUCT.PERFORMANCE_MONITORING, PRODUCT.SESSION_REPLAY]}
-      />
-      {isLoading ? (
-        <LoadingIndicator />
-      ) : isError ? (
-        <LoadingError
-          message={t('Failed to load documentation for the %s platform.', platformName)}
-          onRetry={refetch}
-        />
-      ) : (
-        getDynamicText({
-          value: (
-            <DocsWrapper>
-              <DocumentationWrapper
-                dangerouslySetInnerHTML={{__html: data?.html ?? ''}}
-              />
-              <MissingExampleWarning
-                platform={currentPlatform}
-                platformDocs={{
-                  html: data?.html ?? '',
-                  link: data?.link ?? '',
-                }}
-              />
-            </DocsWrapper>
-          ),
-          fixed: (
-            <Alert type="warning">
-              Platform documentation is not rendered in for tests in CI
-            </Alert>
-          ),
-        })
-      )}
-    </Fragment>
-  );
-}
-
-function ProjectDocs(props: {
-  hasError: boolean;
-  onRetry: () => void;
-  organization: Organization;
-  platform: PlatformKey | null;
-  platformDocs: PlatformDoc | null;
-  project: Project;
-}) {
-  const currentPlatform = props.platform ?? props.project?.platform ?? 'other';
-
-  return (
-    <Fragment>
-      <SetupIntroduction
-        stepHeaderText={t(
-          'Configure %s SDK',
-          platforms.find(p => p.id === currentPlatform)?.name ?? ''
-        )}
-        platform={currentPlatform}
-      />
-      {getDynamicText({
-        value: !props.hasError ? (
-          props.platformDocs !== null && (
-            <DocsWrapper key={props.platformDocs.html}>
-              <DocumentationWrapper
-                dangerouslySetInnerHTML={{__html: props.platformDocs.html}}
-              />
-              <MissingExampleWarning
-                platform={props.platform}
-                platformDocs={props.platformDocs}
-              />
-            </DocsWrapper>
-          )
-        ) : (
-          <LoadingError
-            message={t(
-              'Failed to load documentation for the %s platform.',
-              props.project?.platform
-            )}
-            onRetry={props.onRetry}
-          />
-        ),
-        fixed: (
-          <Alert type="warning">
-            Platform documentation is not rendered in for tests in CI
-          </Alert>
-        ),
-      })}
-    </Fragment>
-  );
-}
-
-function SetupDocs({route, router, location, recentCreatedProject: project}: StepProps) {
+  recentCreatedProject: project,
+}: StepProps) {
   const api = useApi();
   const organization = useOrganization();
 
@@ -294,6 +128,13 @@ function SetupDocs({route, router, location, recentCreatedProject: project}: Ste
   ]);
 
   useEffect(() => {
+    // if project not found, redirect to platform selection
+    if (!project) {
+      router.replace(normalizeUrl(`/onboarding/${organization.slug}/select-platform/`));
+    }
+  }, [project, router, organization.slug]);
+
+  useEffect(() => {
     fetchData();
   }, [fetchData, location.query.product, project?.platform]);
 
@@ -321,7 +162,7 @@ function SetupDocs({route, router, location, recentCreatedProject: project}: Ste
               }}
             />
           ) : showDocsWithProductSelection ? (
-            <DocWithProductSelection
+            <SdkDocWithProductSelection
               organization={organization}
               projectSlug={project.slug}
               location={location}
@@ -337,7 +178,7 @@ function SetupDocs({route, router, location, recentCreatedProject: project}: Ste
                 )}
                 platform={currentPlatform}
               />
-              <SetupDocsLoader
+              <SdkLoader
                 organization={organization}
                 project={project}
                 location={location}
@@ -346,7 +187,7 @@ function SetupDocs({route, router, location, recentCreatedProject: project}: Ste
               />
             </Fragment>
           ) : (
-            <ProjectDocs
+            <SdkDoc
               platform={loadedPlatform}
               project={project}
               hasError={hasError}
@@ -413,8 +254,6 @@ function SetupDocs({route, router, location, recentCreatedProject: project}: Ste
   );
 }
 
-export default SetupDocs;
-
 const AnimatedContentWrapper = styled(motion.div)`
   overflow: hidden;
 `;
@@ -431,7 +270,7 @@ AnimatedContentWrapper.defaultProps = {
   },
 };
 
-const DocsWrapper = styled(motion.div)``;
+export const DocsWrapper = styled(motion.div)``;
 
 DocsWrapper.defaultProps = {
   initial: {opacity: 0, y: 40},
